DUOS-2168 moved consoleTypes to DarCollectionCellData, deleted dar_table components
