+++ src/pages/ResearcherConsole.js
@@ -7,7 +7,7 @@ import { DarCollectionTableColumnOptions, DarCollectionTable } from '../componen
 import accessIcon from '../images/icon_access.png';
 import {Notifications, searchOnFilteredList, getSearchFilterFunctions } from '../libs/utils';
 import SearchBar from '../components/SearchBar';
-import { consoleTypes } from '../components/dar_table/DarTableActions';
+import { consoleTypes } from '../components/dar_collection_table/DarCollectionTableCellData';
 import { USER_ROLES } from '../libs/utils';
 
 const filterFn = getSearchFilterFunctions().darCollections;
